Reorganize merge operation scripts
